+++ Realm/Tests/MigrationTests.mm
@@ -90,7 +90,7 @@
 
 - (void)testSchemaVersion {
     [RLMRealm setDefaultRealmSchemaVersion:1 withMigrationBlock:^(__unused RLMMigration *migration,
-                                                      __unused NSUInteger oldSchemaVersion) {
+                                                                  __unused NSUInteger oldSchemaVersion) {
     }];
 
     RLMRealm *defaultRealm = [RLMRealm defaultRealm];
@@ -105,7 +105,8 @@
 
     XCTAssertEqual(0U, [RLMRealm schemaVersionAtPath:RLMRealm.defaultRealmPath encryptionKey:nil error:nil]);
     [RLMRealm setDefaultRealmSchemaVersion:1 withMigrationBlock:^(__unused RLMMigration *migration,
-                                                                  __unused NSUInteger oldSchemaVersion) {
+                                                                  NSUInteger oldSchemaVersion) {
+        XCTAssertEqual(0U, oldSchemaVersion);
     }];
 
     RLMRealm *realm = [RLMRealm defaultRealm];
@@ -131,7 +132,8 @@
 
     __block bool migrationComplete = false;
     [RLMRealm setSchemaVersion:2 forRealmAtPath:RLMTestRealmPath() withMigrationBlock:^(__unused RLMMigration *migration,
-                                                                                        __unused NSUInteger oldSchemaVersion) {
+                                                                                        NSUInteger oldSchemaVersion) {
+        XCTAssertEqual(0U, oldSchemaVersion);
         migrationComplete = true;
     }];
     RLMRealm *anotherRealm = [RLMRealm realmWithPath:RLMTestRealmPath()];
@@ -701,6 +703,15 @@
     }
 
     RLMClearAccessorCache();
+}
+
+- (void)testMigrationBlockNotCalledForIntialRealmCreation {
+    [RLMRealm setSchemaVersion:1
+                forRealmAtPath:RLMTestRealmPath()
+            withMigrationBlock:^(__unused RLMMigration *migration, __unused NSUInteger oldSchemaVersion) {
+                XCTFail(@"Migration block should not have been called");
+            }];
+    XCTAssertNoThrow([self realmWithTestPath]);
 }
 
 @end